Refactored Viewer and Editor for better feature separation.
