+++ suplemon/editor.py
@@ -51,8 +51,6 @@ class Editor(Viewer):
         """
         Viewer.__init__(self, app, window)
 
-        # Copy/paste buffer
-        self.buffer = []
         # Last search used in 'find'
         self.last_find = ""
         # History of editor states for undo/redo
@@ -62,9 +60,9 @@ class Editor(Viewer):
         # Last editor action that was used (for undo/redo)
         self.last_action = None
 
-        self.operations = {
-            "home": self.home,                            # Home
-            "end": self.end,                              # End
+    def init(self):
+        Viewer.init(self)
+        operations = {
             "backspace": self.backspace,                  # Backspace
             "delete": self.delete,                        # Delete
             "insert": self.insert,                        # Insert
@@ -72,10 +70,6 @@ class Editor(Viewer):
             "tab": self.tab,                              # Tab
             "untab": self.untab,                          # Shift + Tab
             "escape": self.escape,                        # Escape
-            "arrow_right": self.arrow_right,              # Arrow Right
-            "arrow_left": self.arrow_left,                # Arrow Left
-            "arrow_up": self.arrow_up,                    # Arrow Up
-            "arrow_down": self.arrow_down,                # Arrow Down
             "new_cursor_up": self.new_cursor_up,          # Alt + Up
             "new_cursor_down": self.new_cursor_down,      # Alt + Down
             "new_cursor_left": self.new_cursor_left,      # Alt + Left
@@ -94,25 +88,9 @@ class Editor(Viewer):
             "duplicate_line": self.duplicate_line,        # Ctrl + W
             "find_next": self.find_next,                  # Ctrl + D
             "find_all": self.find_all,                    # Ctrl + A
-            "jump_left": self.jump_left,                  # Ctrl + Left
-            "jump_right": self.jump_right,                # Ctrl + Right
-            "jump_up": self.jump_up,                      # Ctrl + Up
-            "jump_down": self.jump_down,                  # Ctrl + Down
         }
-
-    def get_buffer(self):
-        """Returns the current buffer.
-
-        Returns the local buffer or the global buffer depending on config.
-        """
-        if self.app.config["editor"]["use_global_buffer"]:
-            return self.app.global_buffer
-        else:
-            return self.buffer
-
-    def get_key_bindings(self):
-        """Get list of editor key bindings."""
-        return self.app.get_key_bindings()
+        for key in operations.keys():
+            self.operations[key] = operations[key]
 
     def set_buffer(self, buffer):
         """Sets local or global buffer depending on config."""
@@ -176,6 +154,19 @@ class Editor(Viewer):
         self.current_state = index
         self.refresh()
 
+    def handle_input(self, event):
+        done = Viewer.handle_input(self, event)
+        if not done:
+            key = event.key_code
+            name = event.key_name
+            if isinstance(key, str):
+                self.type(key)
+                return True
+            elif name and not name.startswith("KEY_"):
+                self.type(name)
+                return True
+        return False
+
     def undo(self):
         """Undo the last command or change."""
         self.last_action = "undo"
@@ -192,102 +183,6 @@ class Editor(Viewer):
     #
     # Cursor operations
     #
-
-    def arrow_right(self):
-        """Move cursors right."""
-        for cursor in self.cursors:
-            line = self.lines[cursor.y]
-            if cursor.y != len(self.lines)-1 and (cursor.x >= len(line) or len(line) == 0):
-                cursor.move_down()
-                cursor.set_x(0)
-            elif cursor.x < len(self.lines[cursor.y]) and len(line) > 0:
-                cursor.move_right()
-        self.move_cursors()
-        self.scroll_down()
-
-    def arrow_left(self):
-        """Move cursors left."""
-        for cursor in self.cursors:
-            if cursor.y != 0 and cursor.x == 0:
-                cursor.move_up()
-                cursor.set_x(len(self.lines[cursor.y])+1)
-        self.move_cursors((-1, 0))
-        self.scroll_up()
-
-    def arrow_up(self):
-        """Move cursors up."""
-        self.move_cursors((0, -1))
-        self.scroll_up()
-
-    def arrow_down(self):
-        """Move cursors down."""
-        self.move_cursors((0, 1))
-        self.scroll_down()
-
-    def jump_left(self):
-        """Jump one 'word' to the left."""
-        chars = self.config["punctuation"]
-        for cursor in self.cursors:
-            line = self.lines[cursor.y]
-            if cursor.x == 0:
-                if cursor.y > 0:
-                    # Jump to end of previous line
-                    cursor.set_x(len(self.lines[cursor.y-1]))
-                    cursor.move_up()
-                continue
-            if cursor.x <= len(line):
-                cur_chr = line[cursor.x-1]
-            else:
-                cur_chr = line[cursor.x]
-            while cursor.x > 0:
-                next = cursor.x-2
-                if next < 0:
-                    next = 0
-                if cur_chr == " ":
-                    cursor.move_left()
-                    if line[next] != " ":
-                        break
-                else:
-                    cursor.move_left()
-                    if line[next] in chars:
-                        break
-        self.move_cursors()
-
-    def jump_right(self):
-        """Jump one 'word' to the right."""
-        chars = self.config["punctuation"]
-        for cursor in self.cursors:
-            line = self.lines[cursor.y]
-            if cursor.x == len(line):
-                if cursor.y < len(self.lines):
-                    # Jump to start of next line
-                    cursor.set_x(0)
-                    cursor.move_down()
-                continue
-            cur_chr = line[cursor.x]
-            while cursor.x < len(line):
-                next = cursor.x+1
-                if next == len(line):
-                    next -= 1
-                if cur_chr == " ":
-                    cursor.move_right()
-                    if line[next] != " ":
-                        break
-                else:
-                    cursor.move_right()
-                    if line[next] in chars:
-                        break
-        self.move_cursors()
-
-    def jump_up(self):
-        """Jump up 3 lines."""
-        self.move_cursors((0, -3))
-        self.scroll_up()
-
-    def jump_down(self):
-        """Jump down 3 lines."""
-        self.move_cursors((0, 3))
-        self.scroll_down()
 
     def new_cursor_up(self):
         """Add a new cursor one line up."""
@@ -343,34 +238,6 @@ class Editor(Viewer):
         self.cursors = [self.cursors[0]]
         self.move_cursors()
         self.render()
-
-    def page_up(self):
-        """Move half a page up."""
-        amount = int(self.get_size()[1]/2) * -1
-        self.move_cursors((0, amount))
-        self.scroll_up()
-
-    def page_down(self):
-        """Move half a page down."""
-        amount = int(self.get_size()[1]/2)
-        self.move_cursors((0, amount))
-        self.scroll_down()
-
-    def home(self):
-        """Move to start of line or text on that line."""
-        for cursor in self.cursors:
-            wspace = helpers.whitespace(self.lines[cursor.y])
-            if cursor.x == wspace:
-                cursor.set_x(0)
-            else:
-                cursor.set_x(wspace)
-        self.move_cursors()
-
-    def end(self):
-        """Move to end of line."""
-        for cursor in self.cursors:
-            cursor.set_x(len(self.lines[cursor.y]))
-        self.move_cursors()
 
     #
     # Text editing operations
@@ -682,6 +549,12 @@ class Editor(Viewer):
         self.scroll_to_line(cur.y)
         self.move_cursors()
 
+    def find_query(self):
+        """Find in file via user input."""
+        what = self.app.ui.query("Find:", self.last_find)
+        if what:
+            self.find(what)
+
     def find(self, what, findall=False):
         """Find what in data (from top to bottom). Adds a cursor when found."""
         # Sorry for this colossal function
@@ -777,41 +650,6 @@ class Editor(Viewer):
             self.move_y_cursors(cursor.y, 1)
         self.move_cursors()
         self.store_action_state("duplicate_line")
-
-    def handle_input(self, event):
-        """Handle input."""
-        if event.type == "mouse":
-            return False
-        key = event.key_code
-        name = event.key_name
-        # Try match a key to a method and call it
-
-        key_bindings = self.get_key_bindings()
-        operation = None
-        if key in key_bindings.keys():
-            operation = key_bindings[key]
-        elif name in key_bindings.keys():
-            operation = key_bindings[name]
-        if operation:
-            self.run_operation(operation)
-        # Try to type the key into the editor
-        else:
-            if isinstance(key, str):
-                self.type(key)
-            elif name and not name.startswith("KEY_"):
-                self.type(name)
-        return False
-
-    def run_operation(self, operation):
-        """Run an editor core operation."""
-        if operation in self.operations.keys():
-            cancel = self.app.trigger_event_before(operation)
-            if cancel:
-                return False
-            result = self.operations[operation]()
-            self.app.trigger_event_after(operation)
-            return result
-        return False
 
 
 class PromptEditor(Editor):
